Split start.py into APIRouter modules
Move the FastAPI lifespan, watchdog handler, request schemas and the 15
endpoints out of the 760-line start.py into dedicated modules:

- lifespan.py        lifespan + RedisEventHandler + redis_event_consumer
- schemas.py         request BaseModels
- routers/scan.py    /scan
- routers/process.py /processall /maketorrent /upload /seed
- routers/jobs.py    /cjoblist
- routers/posters.py /set{tmdb,tvdb,imdb}id /setposter{url,dname} + update_poster
- routers/settings.py /setting /setenv
- routers/search.py  /filter
- routers/ws.py      /ws

start.py keeps the unit3dwup.start:app and :main entrypoints and only
assembles the app and includes the routers. Behaviour is unchanged:
endpoints reach shared state through request.app.state instead of a module
global. Same routes, same HTTP methods, same logic.
