+++ unit3dwup/routers/settings.py
@@ -1,0 +1,151 @@
+# -*- coding: utf-8 -*-
+import inspect
+import json
+import os
+
+from fastapi import APIRouter, Request, status
+from fastapi.responses import JSONResponse
+
+from dotenv import load_dotenv, dotenv_values
+
+from unit3dwup.config import get_settings
+from unit3dwup.config import get_logger
+
+from unit3dwup.services.lifespan_service import update_mounted_paths
+
+from unit3dwup.schemas import SetEnvRequest
+
+router = APIRouter()
+
+
+@router.post("/setting")
+async def configuration(request: Request):
+    """
+    Load setting from the local configuration file
+
+    Required
+    - job_id is fixed to zero
+
+    Return
+    - none
+    """
+
+    app = request.app
+
+    # Logger
+    frame = inspect.currentframe()
+    logger = get_logger(frame.f_code.co_name)
+
+    # Load json settings and return it to the client
+    job_data = await app.state.job.get_job(job_id='0')
+
+    # /// Check main paths
+    if app.state.settings.prefs.SCAN_PATH == os.getcwd():
+        logger.warning("SCAN PATHS NO SET")
+        await app.state.ws_manager.broadcast({
+            "type": "log",
+            "level": "warn",
+            "message": f"{frame.f_code.co_name} Scan Path not set",
+        })
+    if app.state.settings.prefs.TORRENT_ARCHIVE_PATH == os.getcwd():
+        logger.warning("TORRENT ARCHIVE PATH NOT SET")
+        await app.state.ws_manager.broadcast({
+            "type": "log",
+            "level": "warn",
+            "message": f"{frame.f_code.co_name} Torrent archive path not set",
+        })
+    if os.getcwd() in [app.state.settings.prefs.WATCHER_DESTINATION_PATH, app.state.settings.prefs.WATCHER_PATH]:
+        logger.warning("WATCHER PATHS NO SET")
+        await app.state.ws_manager.broadcast({
+            "type": "log",
+            "level": "warn",
+            "message": f"{frame.f_code.co_name} Watcher Paths not set",
+        })
+
+    # Get data
+    user_prefs = json.loads(job_data)
+
+    return JSONResponse(
+        status_code=status.HTTP_200_OK,
+        content={"userPreferences": user_prefs}
+    )
+
+
+@router.post("/setenv")
+async def set_env(payload: SetEnvRequest, request: Request):
+    """
+    Set the environment variables from the frontend
+    User need to restart the docker
+
+    Required
+    - Key: dictionary key
+    - Value: new value for the environment variable
+
+    Return
+    - none
+    """
+
+    app = request.app
+
+    # Load env file
+    load_dotenv(dotenv_path=app.state.env_file, override=True)
+
+    # Edit file env
+    # Replace set_key in python dotenv
+    # Permission denied when tries to create a temporary file but it fails because the container is not running as root
+    # set_key(str(app.state.env_file), payload.key, payload.value)
+
+    # Current variabiles value
+    env_vars = dotenv_values(str(app.state.env_file))
+    # update the key
+    env_vars[payload.key] = payload.value
+    # Rewrite file env
+    with open(str(app.state.env_file), "w", encoding="utf-8") as f:
+        for k, v in env_vars.items():
+            f.write(f"{k}={v}\n")
+
+    # Refresh memory
+    os.environ[payload.key] = payload.value
+
+    # Clear cache and reload
+    get_settings.cache_clear()
+
+    # Refresh the lru cache
+    settings = get_settings()
+
+    # Update the fast api app state
+    app.state.settings = settings
+
+    # This key requires a Docker restart
+    if payload.key.upper() in ["PREFS__WATCHER_DESTINATION_PATH", "PREFS__WATCHER_PATH", "PREFS__TORRENT_ARCHIVE_PATH",
+                               "PREFS__SCAN_PATH"] and os.getenv("DOCKER") == "true":
+        app.state.restart_docker = True
+
+    # Update mounted path strings. Changes paths require restarting docker when DOCKER == 1
+    await update_mounted_paths(app=app)
+
+    # Ricreate profile (overwrite -> it uses hset command)
+    await app.state.job.create_profile(dict(settings.prefs))
+
+    # Logger
+    frame = inspect.currentframe()
+    logger = get_logger(frame.f_code.co_name)
+
+    # Console message
+    logger.info(f"-> Update {payload.key} value -> {payload.value}\n")
+
+    # Send log to the client
+    await app.state.ws_manager.broadcast({
+        "type": "log",
+        "level": "success",
+        "message": f"-> Update {payload.key} value -> {payload.value}\n",
+    })
+
+    return JSONResponse(
+        status_code=status.HTTP_200_OK,
+        content={
+            "source": "local",
+            "docker": os.getenv("DOCKER"),
+            "message": f"Saved {payload.key}",
+        }
+    )
